# docs/swagger.yaml
basePath: /
definitions:
  github_com_emzola_issuetracker_internal_model.Issue:
    properties:
      actual_resolution_date:
        type: string
      assigned_to:
        type: integer
      created_by:
        type: string
      created_on:
        type: string
      description:
        type: string
      id:
        type: integer
      modified_by:
        type: string
      modified_on:
        type: string
      priority:
        type: string
      progress:
        type: string
      project_id:
        type: integer
      reported_date:
        type: string
      reporter_id:
        type: integer
      resolution_summary:
        type: string
      status:
        type: string
      target_resolution_date:
        type: string
      title:
        type: string
    type: object
...
  internal_handler_http.createIssuePayload:
    properties:
      assigned_to:
        type: integer
      description:
        type: string
      priority:
        type: string
      project_id:
        type: integer
      target_resolution_date:
        type: string
      title:
        type: string
    type: object
...
  internal_handler_http.updateIsssuePayload:
    properties:
      actual_resolution_date:
        type: string
      assigned_to:
        type: integer
      description:
        type: string
      priority:
        type: string
      progress:
        type: string
      resolution_summary:
        type: string
      status:
        type: string
      target_resolution_date:
        type: string
      title:
        type: string
    type: object
...
host: https://issuetracker-api-dev.fl0.io
info:
  contact:
    email: emma.idika@yahoo.com
    name: API Support
  description: This is an API service for an issue tracker.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Issue Tracker API
  version: 1.0.0
paths:
  /v1/issues:
    get:
      description: This endpoint gets all issues
      parameters:
      - description: Bearer token
        in: header
        name: token
        required: true
        type: string
      - description: Query string param for title
        in: query
        name: title
        type: string
      - description: Query string param for reported_date
        in: query
        name: reported_date
        type: string
      - description: Query string param for project_id
        in: query
        name: project_id
        type: string
      - description: Query string param for assigned_to
        in: query
        name: assigned_to
        type: string
      - description: Query string param for status
        in: query
        name: status
        type: string
      - description: Query string param for priority
        in: query
        name: priority
        type: string
      - description: Query string param for pagination (min 1)
        in: query
        name: page
        type: string
      - description: Query string param for pagination (max 100)
        in: query
        name: page_size
        type: string
      - description: 'Sort by asc or desc order. Asc: id, title, reported_date, project_id,
          assigned_to, status, priority | Desc: -id, -title, -reported_date, -project_id,
          -assigned_to, -status, -priority'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_emzola_issuetracker_internal_model.Issue'
            type: array
        "422":
          description: Unprocessable Entity
        "500":
          description: Internal Server Error
      summary: Get all issues
      tags:
      - issues
    post:
      consumes:
      - application/json
      description: Create a new issue with the request payload
      parameters:
      - description: Bearer token
        in: header
        name: token
        required: true
        type: string
      - description: Request payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_handler_http.createIssuePayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_emzola_issuetracker_internal_model.Issue'
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "422":
          description: Unprocessable Entity
        "500":
          description: Internal Server Error
      summary: Create a new issue
      tags:
      - issues
  /v1/issues/{issue_id}:
    delete:
      description: This endpoint deletes an issue
      parameters:
      - description: Bearer token
        in: header
        name: token
        required: true
        type: string
      - description: ID of issue to delete
        in: path
        name: issue_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Delete an issue
      tags:
      - issues
    get:
      description: This endpoint gets an issue by ID
      parameters:
      - description: Bearer token
        in: header
        name: token
        required: true
        type: string
      - description: ID of issue to get
        in: path
        name: issue_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_emzola_issuetracker_internal_model.Issue'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get issue by ID
      tags:
      - issues
    patch:
      consumes:
      - application/json
      description: This endpoint updates an issue
      parameters:
      - description: Bearer token
        in: header
        name: token
        required: true
        type: string
      - description: Request payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_handler_http.updateIsssuePayload'
      - description: ID of issue to update
        in: path
        name: issue_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_emzola_issuetracker_internal_model.Issue'
        "400":
          description: Bad Request
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "409":
          description: Conflict
        "422":
          description: Unprocessable Entity
        "500":
          description: Internal Server Error
      summary: Update an issue
      tags:
      - issues
...
swagger: "2.0"